+++ frontend/src/views/BillingSecurity.vue
@@ -86,6 +86,18 @@
                 </div>
             </div>
             
+            <div v-if="selectedMonthlyPricing || selectedYearlyPricing" class="form-group mb-3">
+                <p class="text-start mb-1" style="font-family: radley; font-size: 25px;">Apply Coupon Code (Leave Blank If Not Applicable)</p>
+                <div class="input-group mb-3">
+                    <input type="text" class="form-control" :style="'border-color: black;'" v-model="couponCode">
+                    <button class="btn btn-outline-secondary" type="button" id="couponButton" @click="applyCoupon()">Apply</button>
+                </div>
+                <span v-if="isFullDiscount" class="text-success">This coupon enables you to subscribe for free! No payment required!</span>
+                <span v-if="paymentAmount > 0" class="text-success">The payment for this subscription will be ${{ paymentAmount/100 }}</span>
+                <span v-if="invalidCoupon" class="text-danger">This coupon is invalid, please try another coupon code</span>
+                <span v-if="failedStripe" class="text-danger">Payment rendering has failed, please try with a different code. Contact administrator if problem persists.</span>
+            </div>
+
             <div v-if="selectedMonthlyPricing || selectedYearlyPricing">
                 <form id="payment-form">
                     <div id="payment-element">
@@ -161,8 +173,8 @@
         data(){
             return{
                 dataLoaded: false,
-                monthlyPriceId: "price_1Q2wvTILhk1xtKohjMfP1Gr0",
-                yearlyPriceId: "price_1Q2zOvILhk1xtKohtZSFuDeb",
+                monthlyPriceId: "",
+                yearlyPriceId: "",
 
                 // owner details
                 stripe: null,
@@ -183,14 +195,17 @@
                 // plan details
                 selectedMonthlyPricing: false,
                 selectedYearlyPricing: true,
-
                 // stripe details
                 priceId: "",
                 customerId: "",
                 clientSecret: "",
-
+                couponCode: "",
+                invalidCoupon: false,
+                failedStripe: false,
                 paymentFilled: false,
-
+                paymentElementInstance: null,
+                isFullDiscount: false,
+                paymentAmount:null,
                 // username and password
                 usernames: null,
                 username: "",
@@ -204,6 +219,8 @@
             }
         },
         async mounted(){
+            this.monthlyPriceId = process.env.VUE_APP_STRIPE_MONTHLY_PRICE_ID,
+            this.yearlyPriceId = process.env.VUE_APP_STRIPE_YEARLY_PRICE_ID,
 
             this.token = this.$route.query.token;
             
@@ -224,8 +241,8 @@
                     await this.create_customer();
                 }
 
-                await this.create_subscription();
-                this.paymentElement();
+                // await this.create_subscription();
+                // this.paymentElement();
             }
 
             initiateProcess.call(this);
@@ -234,6 +251,11 @@
         },
         methods: {
             async toggleYearlyPricing(){
+                this.invalidCoupon=false;
+                this.failedStripe=false;
+                this.isFullDiscount=false;
+                this.paymentAmount = null
+                this.unmountPayment();
                 if(this.selectedMonthlyPricing && !this.selectedYearlyPricing){
                     this.selectedMonthlyPricing= false
                 }
@@ -241,8 +263,6 @@
                     this.selectedYearlyPricing=false
                 }else{
                     this.selectedYearlyPricing = true
-                    await this.create_subscription()
-                    this.paymentElement()
                 }
             },
             async toggleMonthlyPricing(){
@@ -253,8 +273,35 @@
                     this.selectedMonthlyPricing=false
                 }else{
                     this.selectedMonthlyPricing = true
-                    await this.create_subscription()
-                    this.paymentElement()
+                }
+            },
+            async applyCoupon(){
+                this.invalidCoupon=false;
+                this.failedStripe=false;
+                this.isFullDiscount=false;
+                this.paymentAmount = null;
+                this.unmountPayment();
+                try{
+                    let couponResponseCode = null
+                    const response = await this.create_subscription()
+                    if(response.clientSecret){
+                        couponResponseCode = 200
+                        this.paymentAmount = response.amountDue
+                        this.paymentElement()
+                    }else if(response.isFullDiscount){
+                        this.isFullDiscount = true
+                    }
+                    else if(response.status == 404){
+                        this.invalidCoupon = true;
+                        
+                    }else if(response.status == 400){
+                        this.failedStripe = true;
+                    }
+                    return couponResponseCode
+                }
+                catch(error){
+                    console.error(error);
+                    return null
                 }
             },
             async loadData(){
@@ -335,6 +382,7 @@
             },
 
             async create_subscription() {
+                //Change these Ids to be taken from the env file instead
                 if (this.selectedMonthlyPricing) {
                     this.priceId = this.monthlyPriceId;
                 } else {
@@ -356,13 +404,13 @@
                         {
                             priceId: this.priceId,
                             customerId: this.customerId,
+                            couponCode: this.couponCode
                         }, {
                         headers: {
                             'Content-Type': 'application/json'
                         }
                     });
                     const endTime = new Date().getTime();
-                    
                     console.log(`🔍 Request completed in ${endTime - startTime}ms`);
                     this.displayDebugInfo(response);
                     
@@ -380,13 +428,18 @@
                         });
                         
                         return response.data;
-                    } else {
+                    }else if(response.data.isFullDiscount){
+                       return response.data 
+                    }
+                    else {
                         console.error('🔍 No client_secret in response!', response.data);
                         return null;
                     }
                 } catch (error) {
                     console.error('🔍 Subscription creation error:', error.response ? error.response.data : error);
-                    return null;
+                    console.log(error.response.status)
+                    return error.response;
+
                 }
             },
 
@@ -401,28 +454,60 @@
                 this.elements = stripe.elements({clientSecret, appearance, loader: 'auto', allowPromotionCodes: true});
                 const elements = this.elements;
 
-                const paymentElement = elements.create('payment');
-                paymentElement.mount('#payment-element');
+                this.paymentElementInstance = elements.create('payment');
+                this.paymentElementInstance.mount('#payment-element');
 
-                paymentElement.on('change', (event) => {
+                this.paymentElementInstance.on('change', (event) => {
                     this.paymentFilled = event.complete;
                 });
 
             },
+            async unmountPayment(){
+                if (this.paymentElementInstance) {
+                    this.paymentElementInstance.unmount();
+                    this.paymentElementInstance = null;
+                }
+            },
 
             async processPayment() {
-                console.log('🔍 PAYMENT INFO:', {
-                    clientSecret: this.clientSecret ? this.clientSecret.substring(0, 10) + '...' : 'not set'
-                });
-                const stripe = this.stripe;
-                const elements = this.elements;
-
-                const { error, paymentIntent } = await stripe.confirmPayment({
-                    elements,
-                    redirect: 'if_required' // Prevents automatic redirection
-
-                });
-
+                // Updated this function to handle situations with 100% or coupon discount > subscription amount
+                let error, paymentIntent, fullyDiscountedSubscription;
+                if(!this.isFullDiscount){
+                    console.log('🔍 PAYMENT INFO:', {
+                        clientSecret: this.clientSecret ? this.clientSecret.substring(0, 10) + '...' : 'not set'
+                    });
+                    const stripe = this.stripe;
+                    const elements = this.elements;
+    
+                    ({error, paymentIntent} = await stripe.confirmPayment({
+                        elements,
+                        redirect: 'if_required' // Prevents automatic redirection
+    
+                    }));
+                    
+                }else{
+                    // Submit subscription request for the promo code
+                    // Set a flag to true if nothing wrong
+                    try{
+                        const response = await this.$axios.post(`${process.env.VUE_APP_API_URL}/payment/create-full-discount-subscription`,
+                            {
+                                priceId: this.priceId,
+                                customerId: this.customerId,
+                                couponCode: this.couponCode
+                            }, {
+                            headers: {
+                                'Content-Type': 'application/json'
+                            }
+                        });
+                        if(response.data.subscriptionId){
+                            fullyDiscountedSubscription = true
+                        }
+                    }
+                    catch(error){
+                        console.error(error)
+                    }
+                    
+                }
                 if (error) {
                     if (error.type === "card_error" || error.type === "validation_error") {
                         console.log(error.message);
@@ -430,7 +515,7 @@
                         console.log("An unexpected error occurred.");
                     }
 
-                } else if (paymentIntent && paymentIntent.status === 'succeeded') {
+                } else if ((paymentIntent && paymentIntent.status === 'succeeded')|| fullyDiscountedSubscription) {
                     // sucessful payment -> business is now verified
 
                     // update account request
@@ -496,7 +581,7 @@
 
             async createAccount () {
                 
-                if (!this.paymentFilled) {
+                if ((!this.paymentFilled) && (!this.isFullDiscount)) {
                     return;
                 }
                 console.log("creating account");